servers/exchange/requirements.txt
python-dotenv
simplefix
requests
python-socketio[client]
